+++ src/lib.rs
@@ -3,9 +3,11 @@ mod error;
 mod letters_boxed;
 mod prepare_words;
 mod shape;
+mod solution;
 
 pub use cli::{Cli, Commands};
 pub use error::Error;
 pub use letters_boxed::{LettersBoxed, Shuffle};
 pub use prepare_words::PrepareWords;
 pub use shape::Shape;
+pub use solution::Solution;