+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - ✨ add shuffle depth option for word chain(pr [#46])
+- ✨ add solution module for word chain puzzle(pr [#47])
 
 ## [0.1.12] - 2025-04-03
 
@@ -173,6 +174,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [#44]: https://github.com/jerus-org/slb/pull/44
 [#45]: https://github.com/jerus-org/slb/pull/45
 [#46]: https://github.com/jerus-org/slb/pull/46
+[#47]: https://github.com/jerus-org/slb/pull/47
 [Unreleased]: https://github.com/jerus-org/slb/compare/v0.1.12...HEAD
 [0.1.12]: https://github.com/jerus-org/slb/compare/v0.1.11...v0.1.12
 [0.1.11]: https://github.com/jerus-org/slb/compare/v0.1.10...v0.1.11